requirements.txt
python-telegram-bot>=22.7,<23
python-dotenv>=1.2,<2
yt-dlp>=2026.3.17
